agente-notas-fiscais/api-llm/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
langchain
openai
pandas
langchain-community
langchain-experimental
azure-identity
azure-keyvault-secrets
requests
tabulate